Moved tests to account actions
